requirements.txt
markitdown
watchdog
chromadb
mcp
pydantic
PyYAML
click
tqdm
